refactor(toggl): split API implementation from service
